// src/authenticated-app/task/types.ts
/*
 * @Author: zhanghui.chen
 * @Date: 2021-08-04 11:16:22
 * @LastEditTime: 2021-08-04 11:19:00
 * @LastEditors: zhanghui.chen
 */
// 表格类型
export interface TableItemsTypes {
  task_name: string;
  creater: string;
  upload_time: string;
  numbers: string;
  finish: string;
  completed: string;
  undone: string;
  task_type: string;
  remark: string;
}